Update AGO_ItemPropertiesUpdate_fromXLSX.py
This script is great! I noticed in testing that cells left blank in the input Excel file resulted in the corresponding Item Properties being updated with "nan" (i.e. Not A Number). I tested adding a parameter to the pandas.read_excel method to stop this and it worked. The parameter is na_filter=False. In short, because we don't care about blank values in this workflow, we don't need to flag them.

I also noticed a typo: I am pretty sure the initial comments are from the other script that Wes created and vice versa, so I switched them.